+++ packages/liferay-theme-tasks/lib/lookup.js
@@ -26,9 +26,6 @@ function lookup(key, version = null) {
 		baseThemeDependencies: () =>
 			require('./lookup/base').getBaseDependencies,
 		devDependencies: () => require('./lookup/dependencies').devDependencies,
-		'kickstart:choices': () => require('./lookup/kickstart').choices,
-		'kickstart:afterPromptThemeSource': () =>
-			require('./lookup/kickstart').afterPromptThemeSource,
 	}[key]()(version);
 }
